encrypted text saved in file with invr code map and encrypted text

+++ src/main/java/Symbol.kt
@@ -1,6 +1,6 @@
 package main.java
 
-class Symbol(val symbol: String, val freq: Int) : Comparable<Symbol> {
+class Symbol(val symbol: String, val freq: Long) : Comparable<Symbol> {
 
     override fun compareTo(other: Symbol): Int {
         return freq.compareTo(other.freq)
